Move zsl to sam
