+++ tests/unit/algo/visual_prompting/backbones/__init__.py
@@ -1,3 +1,3 @@
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 """Test of custom backbones of OTX visual prompting task."""
