+++ tests/unit/algo/visual_prompting/encoders/__init__.py
@@ -1,3 +1,3 @@
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 """Test of custom encoders of OTX visual prompting task."""
