+++ tests/unit/algo/visual_prompting/losses/__init__.py
@@ -1,3 +1,3 @@
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
-"""Test of custom decoders of OTX visual prompting task."""
+"""Test of custom losses of OTX visual prompting task."""
