+++ tests/unit/algo/visual_prompting/visual_prompters/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+"""Test of custom visual prompters of OTX visual prompting task."""
